refactor: was added lyric table, and date_time column in almost all tables, implemented dashboard on UI, sign in without refresh page, refactored and simplified code

+++ src/main/resources/static/js/player/history/history.js
@@ -2,15 +2,17 @@ let historyList = null;
 let historyTrackButton = $("#historyTrack");
 
 function addTrackToHistory(fullTitleTrack) {
-    $.post({
-        url: '/history',
-        data: {
-            fullTrackTitle: fullTitleTrack
-        },
-        success: function () {
-            console.log(fullTitleTrack + " - added to history");
-        }
-    });
+    if (currentUserUsername.text() !== "") {
+        $.post({
+            url: '/history',
+            data: {
+                fullTrackTitle: fullTitleTrack
+            },
+            success: function () {
+                console.log(fullTitleTrack + " - added to history");
+            }
+        });
+    }
 }
 
 historyTrackButton.on("click", function () {
@@ -21,6 +23,7 @@ historyTrackButton.on("click", function () {
 });
 
 function getHistory() {
+    getLoader();
     listOfTrack = null;
 
     let html = '';
@@ -29,41 +32,45 @@ function getHistory() {
             historyList = data;
             listOfTrack = data;
             countOfTrack = data.length;
+            html += '<h2>History</h2>';
+            html += '<div style="display: grid; grid-template-columns: 200px 200px 200px 200px 200px 175px; grid-auto-rows: 250px;">';
             $.each(data, function (i, track) {
-                html += '<tr>';
-                html += '<td class="commonTd" style="display: flex">';
-                html += '<img id="cover" style="width: 74px; height: 74px; margin-right: 3%" src="data:image/jpeg;charset=utf-8;base64,' + track.byteOfPicture + '"/>';
-                html += '<div class="titleOfTrackInTable">' + track.fullTitle + '</div>';
-                html += '<div class="menuElement" onclick="removeTrackFromHistory(' + track.id + ')">Remove</div>';
-                html += '</td>';
-                html += '</tr>';
+                html += '<div style="max-width: 173px;">';
+                html += '   <div id="musicId" hidden>' + track.id + '</div>';
+                html += '   <img id="cover" style="width: 173px; height: 173px;" src="data:image/jpeg;charset=utf-8;base64,' + track.byteOfPicture + '"/>';
+                html += '   <div class="titleOfTrackInTable" style="text-overflow: ellipsis; white-space: nowrap; overflow: hidden; height: auto">' + track.fullTitle + '</div>';
+                html += '   <div class="menuElement" onclick="removeTrackFromHistory(' + track.id + ')">Remove</div>';
+                html += '</div>'
             });
-            $('#mainTableTBody').html(html);
-            $('#listOfTrack').removeClass("main");
+            supportDashboard.html(html);
+            closeLoader();
         } else {
-            $("#homeButton").notify("You not play any tracks", {
-                position: 'bottom left',
+            closeLoader();
+            $.notify("You not play any tracks", {
+                position: 'top left',
                 className: 'error'
             });
-            getList(lastList);
+            getList(lastUrl);
         }
     });
 }
 
 function removeTrackFromHistory(trackId) {
-    $.ajax({
-        type: 'DELETE',
-        url: '/history',
-        data: {
-            trackId: trackId
-        },
-        success: function (fullTitle) {
-            $.notify(fullTitle + " deleted from your history", {
-                position: 'top right',
-                className: 'info'
-            });
-            console.log(fullTitle + ", deleted from history");
-            getHistory();
-        }
-    });
+    if (currentUserUsername.text() !== "") {
+        $.ajax({
+            type: 'DELETE',
+            url: '/history',
+            data: {
+                trackId: trackId
+            },
+            success: function (fullTitle) {
+                $.notify(fullTitle + " deleted from your history", {
+                    position: 'top left',
+                    className: 'info'
+                });
+                console.log(fullTitle + ", deleted from history");
+                getHistory();
+            }
+        });
+    }
 }
